+++ src/ArduinoJson/Strings/ArduinoStringAdapter.hpp
@@ -39,6 +39,10 @@ class ArduinoStringAdapter {
     return _str->length();
   }
 
+  const char* begin() const {
+    return _str->c_str();
+  }
+
   typedef storage_policies::store_by_copy storage_policy;
 
  private: