+++ src/ArduinoJson/Strings/ConstRamStringAdapter.hpp
@@ -39,6 +39,10 @@ class ConstRamStringAdapter {
     return _str;
   }
 
+  const char* begin() const {
+    return _str;
+  }
+
   typedef storage_policies::store_by_address storage_policy;
 
  protected: